backend/figma2html/requirements.txt
# Core dependencies
requests>=2.28.0
click>=8.0.0
colorama>=0.4.4

# Optional dependencies for enhanced functionality
python-dotenv>=0.19.0
aiohttp>=3.8.0
